Functions working?

+++ vm/include/lit_object.h
@@ -107,7 +107,7 @@ LitClass* lit_new_class(LitVm* vm, LitString* name, LitClass* super);
 LitInstance* lit_new_instance(LitVm* vm, LitClass* klass);
 
 LitString* lit_make_string(LitVm* vm, char* chars, int length);
-LitString* lit_copy_string(LitVm* vm, const char* chars, int length);
+LitString* lit_copy_string(LitVm* vm, const char* chars, size_t length);
 
 static inline bool lit_is_object_type(LitValue value, LitObjectType type) {
 	return IS_OBJECT(value) && AS_OBJECT(value)->type == type;
